requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==2.2.2
numpy==1.14.5
opencv-python==3.4.3.18
Paste==3.1.0
Pillow==5.2.0
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.12
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
waitress==1.3.0
Werkzeug==0.15.5
